requirements.txt

fastapi
uvicorn
pydantic
pyyaml
